% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAE-data.R
\docType{data}
\name{CCAE_perform}
\alias{CCAE_perform}
\title{CCAE model test/train performance}
\format{An object of class \code{list} of length 7.}
\usage{
data(CCAE_perform)
}
\description{
Data containing the performance info for the test/train sets
during model development
}
\examples{
data(CCAE_perform)
}
\keyword{datasets}

