% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractData}
\alias{extractData}
\title{extractData}
\usage{
extractData(connectionDetails, cdmDatabaseSchema, targetDatabaseSchema,
  targetCohortTable = "extValCohort", targetCohortId = 1,
  outcomeCohortId = 2)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{targetDatabaseSchema}{The name of the database schema that is the location where the
cohort data used to define the at risk cohort will be written.
Requires read/write permissions to this database.}

\item{targetCohortTable}{The tablename that will contain the at risk cohort and outcome.
The table has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{targetCohortId}{A unique identifier to define the at risk cohort.}

\item{outcomeCohortId}{A cohort_definition_id used to define the outcome.}
}
\value{
A list containing the plpData and population
}
\description{
Extracts the plpData required for plpModel being validated using the same settings as the model
}
\details{
The function runs various sql scripts to create the target population and outcome populations and then
extracts the data from the database and creates the data objects required to run the model and evaluate it
}

