% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCD-data.R
\docType{data}
\name{MDCD_perform}
\alias{MDCD_perform}
\title{MDCD model test/train performance}
\format{An object of class \code{list} of length 7.}
\usage{
data(MDCD_perform)
}
\description{
Data containing the performance info for the test/train sets
during model development
}
\examples{
data(MDCD_perform)
}
\keyword{datasets}

