% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCR-data.R
\docType{data}
\name{MDCR_perform}
\alias{MDCR_perform}
\title{MDCR model test/train performance}
\format{An object of class \code{list} of length 7.}
\usage{
data(MDCR_perform)
}
\description{
Data containing the performance info for the test/train sets
during model development
}
\examples{
data(MDCR_perform)
}
\keyword{datasets}

