% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTUM-data.R
\docType{data}
\name{OPTUM_perform}
\alias{OPTUM_perform}
\title{OPTUM model test/train performance}
\format{An object of class \code{list} of length 7.}
\usage{
data(CCAE_perform)
}
\description{
Data containing the performance info for the test/train sets
during model development
}
\examples{
data(OPTUM_perform)
}
\keyword{datasets}

