% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viewModels}
\alias{viewModels}
\title{applyModels}
\usage{
viewModels()
}
\value{
returns tables for each model in the package displaying the top 10 variables that increased or
decreased the risk of the outcome
}
\description{
This function shows same of the prognostic factors of the various package models
}
\details{
The function presents a view of the existing models
}

