% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viewValidation}
\alias{viewValidation}
\title{viewValidation}
\usage{
viewValidation(outputDir, save = F)
}
\arguments{
\item{outputDir}{The path to the directory where the results were saved}

\item{save}{Binary indicating whether to save the plots in the outputDir}
}
\value{
Returns plots of the validation performance and the test/train performance of the development dataset
}
\description{
This function lets users view the external validation and internal validaiton of each model
}
\details{
The function creates various ROC and calibration plots for viewing the validation performance
}

